\name{utils}
\alias{whoami}
\title{A whoami test function}
\usage{
    whoami()
}
\description{
    A simple test function.
}
\value{The name of the package as a character string} 
\examples{

    myfunction <- function(){
        iam <- FALSE
    }
    myfunction()
    whoami()
    \dontrun{
       # test a dontrun
       a = 1
    }
} 
